Map 4.0 Layout Change
Altered the placement of a few power poles, all the houses, and the right-side road. Also removed a few trees in the way.
There's either an issue with the build/wire-connection radius itself, or the radius outline that you see when hovering over an object.
Not sure why, but Unity changed the EditorUserSettings. Not sure what that means, either.